requirements.txt
# File: requirements.txt
# Description: Python package dependencies for ThreatSim cybersecurity simulation framework
# Purpose: Comprehensive dependency specification for development and production environments
#
# Copyright (c) 2025 ThreatEcho
# Licensed under the GNU General Public License v3.0
# See LICENSE file for details.

# Core Reinforcement Learning Framework
gymnasium>=0.29.0           # OpenAI Gym replacement for RL environments
pettingzoo>=1.24.0          # Multi-agent reinforcement learning environments
stable-baselines3>=2.3.0    # State-of-the-art RL algorithms (PPO, SAC, etc.)

# Scientific Computing and Data Analysis
numpy>=1.24.0               # Numerical computing foundation
pandas>=2.0.0               # Data manipulation and analysis
scipy>=1.10.0               # Scientific computing library for statistical analysis

# Network and Graph Analysis
networkx>=3.0.0             # Network topology modeling and analysis

# Visualization and Plotting
matplotlib>=3.7.0           # Comprehensive plotting library for analysis
seaborn>=0.12.0             # Statistical data visualization

# Configuration and Data Handling
pyyaml>=6.0.0               # YAML configuration file parsing
jsonschema>=4.17.0          # JSON schema validation for configurations

# Machine Learning and AI
scikit-learn>=1.3.0         # Machine learning utilities and metrics
torch>=2.0.0                # PyTorch for advanced neural network models (optional)

# Development and Testing Dependencies
pytest>=7.0.0               # Testing framework
pytest-cov>=4.0.0           # Code coverage reporting
black>=23.0.0               # Code formatting
flake8>=6.0.0               # Code linting
mypy>=1.0.0                 # Static type checking

# Documentation
sphinx>=6.0.0               # Documentation generation
sphinx-rtd-theme>=1.2.0     # ReadTheDocs theme for documentation

# Performance and Optimization
numba>=0.57.0               # JIT compilation for performance-critical code (optional)
psutil>=5.9.0               # System and process utilities for monitoring
